FIX BUG: cell value out of 0 to 255 range
This patch, fixes the problem of setting the value of a cell, out of the
range of (0,256). So, if the cell be set to -1, it gets translated to
255 and if it be set to 256, it gets translated to 0. It can also handle
values less than -1 and more than 256. For example cell value of -3 gets
translated into 253.

There is also one more test included in the test_ahangarha.py which
tests a code which makes a negative cell value.

This hopefully is the last commit in this code :)